refactor validator chain op args
